requirements.txt
pytest>=7.4
selenium>=4.22
allure-pytest>=2.13.5
webdriver-manager>=4.0.2